refactor: Change idAgendamento type from int to long in ILtdbLtvcService

+++ Common/Services/ILtdbLtvcService.cs
@@ -25,7 +25,7 @@ namespace Common.Services
             string? placa,
             string? motorista,
             string? moveType,
-            int? idAgendamento,
+            long? idAgendamento,
             string? status,
             int pageSize,
             int pageNumber);
